examples/requirements.txt
# MAI-UI Examples Requirements
# For running optimized inference with vLLM

# Core inference engine
vllm>=0.6.0

# Image processing
Pillow>=9.0.0

# OpenAI-compatible client (for API mode)
openai>=1.0.0

# Templating for prompts
Jinja2>=3.0.0

# For notebooks
ipython>=8.0.0
jupyter>=1.0.0

# Benchmarking (optional)
tqdm>=4.60.0
tabulate>=0.9.0

# Transformers (vLLM dependency)
transformers>=4.45.0

# PyTorch (will be installed by vLLM)
# torch>=2.0.0

